refactor(libs): extract wrapped text module
