+++ Questie-BCC.toc
@@ -138,6 +138,7 @@ Localization\lookups\TBC\lookupQuests\lookupQuests.xml
 # Libs
 Modules\Libs\DistanceUtils.lua
 Modules\Libs\utf8.lua
+Modules\Libs\WrappedText.lua
 Modules\Libs\QuestieLib.lua
 Modules\Libs\QuestieSerializer.lua
 Modules\Libs\QuestieCombatQueue.lua
