+++ Questie-Cata.toc
@@ -151,6 +151,7 @@ Localization\lookups\Cata\lookupQuests\lookupQuests.xml
 # Libs
 Modules\Libs\DistanceUtils.lua
 Modules\Libs\utf8.lua
+Modules\Libs\WrappedText.lua
 Modules\Libs\QuestieLib.lua
 Modules\Libs\QuestieSerializer.lua
 Modules\Libs\QuestieCombatQueue.lua
