+++ Questie-WOTLKC.toc
@@ -146,6 +146,7 @@ Localization\lookups\Wotlk\lookupQuests\lookupQuests.xml
 # Libs
 Modules\Libs\DistanceUtils.lua
 Modules\Libs\utf8.lua
+Modules\Libs\WrappedText.lua
 Modules\Libs\QuestieLib.lua
 Modules\Libs\QuestieSerializer.lua
 Modules\Libs\QuestieCombatQueue.lua
